Getrlimit not supported by windows
